DHFPROD-4887: Enabling Spring profiles
Tossing SwaggerConfig into dev profile so we have an easy way of omitting it
