+++ marklogic-data-hub-central/build.gradle
@@ -44,6 +44,7 @@ bootJar {
 bootRun.doFirst {
     environment "mlHost", mlHost
     environment "hubUseLocalDefaults", hubUseLocalDefaults
+    environment "spring.profiles.active", springProfiles
 }
 
 repositories {
