+++ marklogic-data-hub-central/gradle.properties
@@ -5,7 +5,9 @@ springBootJarName=marklogic-data-hub-central
 
 springBootVersion=2.2.6.RELEASE
 springSecurityVersion=5.2.3.RELEASE
-springVersion=5.2.5.RELEASE
+
+# Passed to bootRun as the value of spring.profiles.active
+springProfiles=dev
 
 # For testing locally.
 # Each of these can be overridden via e.g. -PmlHost=somehost when running "./gradlew bootRun".
