+++ marklogic-data-hub-central/src/main/resources/application.properties
@@ -1,3 +1,7 @@
+# Controls which Spring Profile values for enabled. This is a handy way to control which beans are active in a given
+# environment. See SwaggerConfig as an example.
+spring.profiles.active=dev
+
 server.port=8080
 spring.thymeleaf.prefix=classpath:/static/
